+++ lib/providers/create_request_screen_providers/selected_request_provider.dart
@@ -3,12 +3,13 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 final selectedLeaveRequestProvider = StateProvider<LeaveRequest>((ref) {
   return LeaveRequest(
-      id: '',
-      user: '',
-      viewPermission: ViewPermission.everyone,
-      startDate: DateTime.now(),
-      endDate: DateTime.now(),
-      reason: '',
-      leaveType: LeaveType.other,
-      status: RequestStatus.pending);
+    id: '',
+    user: '',
+    viewPermission: ViewPermission.everyone,
+    startDate: DateTime.now(),
+    endDate: DateTime.now(),
+    reason: '',
+    leaveType: LeaveType.other,
+    status: RequestStatus.pending,
+  );
 });
