+++ lib/providers/home_screen_providers/filtered_users_list_provider.dart
@@ -3,7 +3,6 @@ import 'package:cobe_hive_mobile_app/providers/network_providers/user_data_provi
 import 'package:cobe_hive_mobile_app/providers/home_screen_providers/user_filter_provider.dart';
 import 'package:cobe_hive_mobile_app/providers/home_screen_providers/user_search_provider.dart';
 import 'package:cobe_hive_mobile_app/widgets/admin_and_home_widgets/chip_list.dart';
-import 'package:flutter/foundation.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 final statusFilterMap = {
@@ -20,16 +19,16 @@ final filteredUserListProvider = FutureProvider<List<User>>((ref) async {
   final searchQuery = ref.watch(userSearchProvider);
   final userList = ref.watch(userDataProvider);
   final userFilter = ref.watch(userFilterProvider);
+
   List<User> filteredUsers = userList.where(
     (user) {
-      debugPrint('userList: ${user.name} ${user.surname} ${user.status}}');
-
       if (userFilter.contains(StatusFilter.all) || userFilter.isEmpty) {
         return true;
       }
       return userFilter.contains(statusFilterMap[user.status]);
     },
   ).toList();
+
   List<User> searchedUsers = filteredUsers.where((user) {
     if (searchQuery.isEmpty) {
       return true;
