+++ lib/screens/create_request_screen.dart
@@ -1,12 +1,10 @@
-import 'package:cobe_hive_mobile_app/data/constants/app_colors.dart';
-import 'package:cobe_hive_mobile_app/providers/create_request_screen_providers/leave_reason_provider.dart';
 import 'package:cobe_hive_mobile_app/widgets/create_request_widgets/create_request_buttons.dart';
 import 'package:cobe_hive_mobile_app/widgets/create_request_widgets/create_request_header.dart';
 import 'package:cobe_hive_mobile_app/widgets/create_request_widgets/date_picker.dart';
+import 'package:cobe_hive_mobile_app/widgets/create_request_widgets/leave_reason_text_field.dart';
 import 'package:cobe_hive_mobile_app/widgets/create_request_widgets/type_of_leave_selector.dart';
 import 'package:cobe_hive_mobile_app/widgets/create_request_widgets/view_permission_selector.dart';
 import 'package:flutter/material.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 class CreateRequestScreen extends StatelessWidget {
   const CreateRequestScreen({super.key});
@@ -37,12 +35,13 @@ class CreateRequestScreen extends StatelessWidget {
                   child: Text('Reason'),
                 ),
               ),
-              _LeaveReasonTextField(),
+              LeaveReasonTextField(),
               Padding(
                 padding: EdgeInsets.symmetric(vertical: 18),
                 child: Align(
-                    alignment: Alignment.centerLeft,
-                    child: Text('Who can view my request?')),
+                  alignment: Alignment.centerLeft,
+                  child: Text('Who can view my request?'),
+                ),
               ),
               ViewPermissionSelector(),
               Expanded(flex: 2, child: SizedBox()),
@@ -51,47 +50,6 @@ class CreateRequestScreen extends StatelessWidget {
             ],
           ),
         ),
-      ),
-    );
-  }
-}
-
-class _LeaveReasonTextField extends ConsumerWidget {
-  const _LeaveReasonTextField();
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return Container(
-      decoration: BoxDecoration(
-          color: Colors.white,
-          borderRadius: BorderRadius.circular(8),
-          boxShadow: const [
-            BoxShadow(
-              color: AppColors.shadow,
-              blurRadius: 16,
-              offset: Offset(0, 5),
-            )
-          ]),
-      child: TextField(
-        maxLines: 4,
-        autofocus: false,
-        decoration: InputDecoration(
-          hintText: 'Type something ...',
-          hintStyle: const TextStyle(
-            color: AppColors.textAccent,
-            fontSize: 18,
-            fontWeight: FontWeight.w100,
-          ),
-          filled: true,
-          fillColor: Colors.white,
-          enabledBorder: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(8),
-            borderSide: const BorderSide(color: Colors.white),
-          ),
-          contentPadding: const EdgeInsets.all(18),
-        ),
-        onChanged: (value) =>
-            ref.watch(leaveReasonProvider.notifier).state = value,
       ),
     );
   }
