+++ lib/widgets/create_request_widgets/create_request_buttons.dart
@@ -1,6 +1,7 @@
 import 'package:cobe_hive_mobile_app/data/constants/app_colors.dart';
 import 'package:cobe_hive_mobile_app/providers/leave_request_providers/leave_request_options_provider.dart';
 import 'package:cobe_hive_mobile_app/providers/leave_request_providers/leave_request_list_provider.dart';
+import 'package:cobe_hive_mobile_app/widgets/create_request_widgets/request_created_alert_dialog.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
@@ -27,11 +28,12 @@ class CreateRequestButtons extends ConsumerWidget {
           child: ElevatedButton(
             onPressed: () {
               showDialog(
-                  context: context,
-                  barrierDismissible: false,
-                  builder: (BuildContext context) {
-                    return const _RequestCreatedAlertDialog();
-                  });
+                context: context,
+                barrierDismissible: false,
+                builder: (BuildContext context) {
+                  return const RequestCreatedAlertDialog();
+                },
+              );
               ref
                   .read(leaveRequestListProvider.notifier)
                   .addLeaveRequest(leaveRequest);
@@ -57,46 +59,5 @@ class CreateRequestButtons extends ConsumerWidget {
         ),
       ],
     );
-  }
-}
-
-class _RequestCreatedAlertDialog extends StatelessWidget {
-  const _RequestCreatedAlertDialog();
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-        title: const Center(child: Text('Request Created')),
-        titleTextStyle: Theme.of(context).textTheme.bodyLarge,
-        titlePadding: const EdgeInsets.only(top: 30),
-        content: const Text(
-          'Your request is created and it is currently being reviewed.',
-          textAlign: TextAlign.center,
-        ),
-        contentTextStyle: Theme.of(context).textTheme.labelLarge,
-        actions: [
-          OutlinedButton(
-            onPressed: () {
-              int count = 0;
-              Navigator.of(context).popUntil((_) => count++ >= 2);
-            },
-            style: OutlinedButton.styleFrom(
-                side: const BorderSide(color: AppColors.primary),
-                shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(8))),
-            child: const Text(
-              'Okay',
-              style: TextStyle(
-                color: AppColors.textOrange,
-                fontSize: 16,
-              ),
-            ),
-          ),
-        ],
-        actionsPadding: const EdgeInsets.only(bottom: 20),
-        actionsAlignment: MainAxisAlignment.center,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(14),
-        ));
   }
 }
